Removed commented self.box functionality. Made self.region a property again.
Self.box is not used anymore, so it's probably better to just remove its related commented code now. As for self.region, the @property decorator permits us to treat it as a variable (so we can call it as self.region instead of self.region()). It will still do the calculations in the function normally. From my tests the screenshots are being captured correctly after these changes.