Reorganized GUI in an attempt to make it look a bit better.
Reorganized button positions with grid instead of pack. Changed buttons to ttk instead of tk (ttk usually looks better by default). Made it so both main window and top_levels appear in the same place in the screen. Also renamed IMG_DIR to lowercase as it is not a global parameter anymore.